Fix ffi arg/return type mismatches
These are wrong on any LLP64 platform (Windows), where long is 32-bit while
size_t/uintptr_t are 64-bit, and the mismatched callbacks are wrong everywhere:

- size_t parameters (LLVMCreateMemoryBufferWithMemoryRange{,Copy},
  LLVMGetBufferSize, LLVMInitializeMCJITCompilerOptions,
  LLVMCreateMCJITCompilerForModule) were declared :ulong.
- LLVMGetGlobalValueAddress/LLVMGetFunctionAddress return uint64_t, not
  unsigned long - truncated to 32 bits on Windows.
- the MCJIT memory manager callbacks had a spurious leading parameter
  (ffi_gen turned the return type into the first argument), returned :uchar
  where the C API returns uint8_t *, and declared uintptr_t as :ulong.

+++ lib/llvm/core_ffi.rb
@@ -5261,7 +5261,7 @@ module LLVM::C
   # @param [Integer] requires_null_terminator
   # @return [FFI::Pointer(MemoryBufferRef)]
   # @scope class
-  attach_function :create_memory_buffer_with_memory_range, :LLVMCreateMemoryBufferWithMemoryRange, [:string, :ulong, :string, :int], :pointer
+  attach_function :create_memory_buffer_with_memory_range, :LLVMCreateMemoryBufferWithMemoryRange, [:string, :size_t, :string, :int], :pointer
 
   # (Not documented)
   #
@@ -5271,7 +5271,7 @@ module LLVM::C
   # @param [String] buffer_name
   # @return [FFI::Pointer(MemoryBufferRef)]
   # @scope class
-  attach_function :create_memory_buffer_with_memory_range_copy, :LLVMCreateMemoryBufferWithMemoryRangeCopy, [:string, :ulong, :string], :pointer
+  attach_function :create_memory_buffer_with_memory_range_copy, :LLVMCreateMemoryBufferWithMemoryRangeCopy, [:string, :size_t, :string], :pointer
 
   # (Not documented)
   #
@@ -5287,7 +5287,7 @@ module LLVM::C
   # @param [FFI::Pointer(MemoryBufferRef)] mem_buf
   # @return [Integer]
   # @scope class
-  attach_function :get_buffer_size, :LLVMGetBufferSize, [:pointer], :ulong
+  attach_function :get_buffer_size, :LLVMGetBufferSize, [:pointer], :size_t
 
   # (Not documented)
   #
